+++ BlogService.Abstractions/BlogService.Abstractions.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\BlogService.DataModels\BlogService.DataModels.csproj" />
+  </ItemGroup>
+
+</Project>
