+++ BlogService.Abstractions/Repositories/ICommentRepository.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using BlogService.Models;
+using BlogService.DataModels;
 
-namespace BlogService.Abstractions
+namespace BlogService.Abstractions.Repositories
 {
     public interface ICommentRepository
     {
