+++ BlogService.DataModels/Comment.cs
@@ -1,33 +1,33 @@
 ﻿using System;
-using System.Collections.Generic;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 
-namespace BlogService.Models
+namespace BlogService.DataModels
 {
-    public class Post
+    public class Comment
     {
         public string Body { get; set; } = string.Empty;
 
         [BsonDateTimeOptions(Kind = DateTimeKind.Local)]
         public DateTime CreatedOn { get; set; } = DateTime.Now;
 
+        public string Email { get; set; } = string.Empty;
 
-        public HeaderImage Header { get; set; } = new HeaderImage();
+        public string FirstName { get; set; } = string.Empty;
 
         //external Id
         public string Id { get; set; }
 
-        [BsonId]
-
         //generated by MongoDb
         public ObjectId InternalId { get; set; }
 
-        public List<string> Tags { get; set; } = new List<string>();
+        public string LastName { get; set; } = string.Empty;
 
-        public string Title { get; set; } = string.Empty;
+        public string PostId { get; set; }
 
         [BsonDateTimeOptions(Kind = DateTimeKind.Local)]
         public DateTime UpdatedOn { get; set; } = DateTime.Now;
+
+        public string Website { get; set; } = string.Empty;
     }
 }
