+++ BlogService.Domain/Repositories/PostRepository.cs
@@ -1,17 +1,16 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
-using BlogService.Abstractions;
+using BlogService.Abstractions.Repositories;
 using BlogService.Contexts;
-using BlogService.Models;
+using BlogService.DataModels;
+using BlogService.Domain.Mongo;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using MongoDB.Bson;
 using MongoDB.Driver;
-using MongoDB.Driver.Linq;
 
-namespace BlogService.Repositories
+namespace BlogService.Domain.Repositories
 {
     public class PostRepository : BaseMongoRepository, IPostRepository
     {
